LayoutTests/security/cocoa/check-enhanced-security-disabled.html

<!DOCTYPE html>
<html>
<head>
<title>Checks Enhanced Security Isn't Enabled</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<script>

test(() => {
    assert_equals(window.internals.webContentProcessVariant(), "standard", `WebViews should be opted out of EnhancedSecurity when tests aren't in enhanced-security/ directories`);
}, "Check WebView is opted out of EnhancedSecurity");

</script>
</body>
</html>
